Move ItemManifest code
