+++ emergence_lib/src/asset_management/manifest/raw.rs
@@ -2,15 +2,13 @@
 //!
 //! The processing will primarily remove the string IDs and replace them by numbers.
 
-use bevy::{reflect::TypeUuid, utils::HashMap};
+use bevy::reflect::TypeUuid;
 use serde::Deserialize;
 
-use crate::items::ItemData;
-
-use super::{Item, Manifest};
+use super::Manifest;
 
 /// A utility trait to ensure that all trait bounds are satisfied.
-pub(crate) trait RawManifest:
+pub trait RawManifest:
     std::fmt::Debug + TypeUuid + Send + Sync + for<'de> Deserialize<'de> + 'static
 {
     /// The marker type for the manifest ID.
@@ -24,48 +22,4 @@ pub(crate) trait RawManifest:
 
     /// Process the raw manifest from the asset file to the manifest data used in-game.
     fn process(&self) -> Manifest<Self::Marker, Self::Data>;
-}
-
-/// The item data as seen in the original manifest file.
-///
-/// This will be converted to [`crate::items::ItemData`].
-#[derive(Debug, Clone, Deserialize)]
-pub struct RawItemData {
-    /// The maximum number of items that can fit in a stack.
-    stack_size: usize,
-}
-
-impl From<&RawItemData> for ItemData {
-    fn from(value: &RawItemData) -> Self {
-        Self::new(value.stack_size)
-    }
-}
-
-/// The item manifest as seen in the manifest file.
-#[derive(Debug, Clone, Deserialize, TypeUuid)]
-#[uuid = "cd9f4571-b0c4-4641-8d27-1c9c5ad4c812"]
-pub(crate) struct RawItemManifest {
-    /// The data for each item.
-    items: HashMap<String, RawItemData>,
-}
-
-impl RawManifest for RawItemManifest {
-    type Marker = Item;
-    type Data = ItemData;
-
-    fn path() -> &'static str {
-        "manifests/items.manifest.json"
-    }
-
-    fn process(&self) -> Manifest<Self::Marker, Self::Data> {
-        let mut manifest = Manifest::new();
-
-        for (name, raw_data) in &self.items {
-            let data = Self::Data::from(raw_data);
-
-            manifest.insert(name, data)
-        }
-
-        manifest
-    }
 }
