+++ emergence_lib/src/signals.rs
@@ -4,6 +4,7 @@
 //! we can scale path-finding and decisionmaking in a clear and comprehensible way.
 
 use crate as emergence_lib;
+use crate::items::item_manifest::{Item, ItemManifest};
 use crate::structures::structure_manifest::{Structure, StructureManifest};
 use crate::units::unit_manifest::{Unit, UnitManifest};
 use bevy::{prelude::*, utils::HashMap};
@@ -12,7 +13,7 @@ use emergence_macros::IterableEnum;
 use itertools::Itertools;
 use rand::seq::SliceRandom;
 
-use crate::asset_management::manifest::{Id, Item, ItemManifest};
+use crate::asset_management::manifest::Id;
 use crate::simulation::geometry::{MapGeometry, TilePos};
 use crate::simulation::SimulationSet;
 use crate::units::goals::Goal;
